requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
pydantic==2.8.2
pymilvus>=2.4.6
sentence-transformers>=2.7.0
numpy>=1.26
python-dotenv>=1.0.1
httpx>=0.27.0

